client/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
grapheme==0.6.0
pycparser==2.22